Set Up Singleton

+++ M.E.K.A Defence/Assets/Prefabs/Enemy/EnemyHealth.cs
@@ -2,8 +2,8 @@
 
 public class EnemyHealth : MonoBehaviour
 {
-    public int startingHealth = 100;
-    public int currentHealth;
+    public float startingHealth = 100;
+    public float currentHealth;
     public float sinkSpeed = 2.5f;
     //public int scoreValue = 10;
     public AudioClip deathClip;
@@ -39,7 +39,7 @@ public class EnemyHealth : MonoBehaviour
 
     //Function that is called when the player deals damage to you
     //Default condition format
-    public void TakeDamage(int amount, Vector3 hitPoint)
+    public void TakeDamage(float amount, Vector3 hitPoint)
     {
         if (isDead)
             return;
